test: extract persistence helpers
