+++ tests/e2e/canvas.persistence.spec.js
@@ -1,7 +1,4 @@
 import { test, expect } from '@playwright/test';
-import fs from 'node:fs/promises';
-import os from 'node:os';
-import path from 'node:path';
 
 import {
     addEntity,
@@ -11,59 +8,12 @@ import {
     getSavedDiagram,
 } from '../helpers/canvas';
 
-async function exportCurrentDiagram(page) {
-    const downloadPromise = page.waitForEvent('download');
-
-    await page.getByRole('button', { name: 'Exportar JSON' }).click();
-
-    const dialog = page.getByRole('dialog');
-    await expect(
-        dialog.getByText('Exportación diagrama en JSON'),
-    ).toBeVisible();
-
-    await dialog.getByRole('button', { name: 'Aceptar' }).click();
-
-    const download = await downloadPromise;
-    const savePath = path.join(
-        os.tmpdir(),
-        `diagram-${Date.now()}-${download.suggestedFilename()}`,
-    );
-
-    await download.saveAs(savePath);
-
-    const raw = await fs.readFile(savePath, 'utf8');
-    return JSON.parse(raw);
-}
-
-async function resetDiagram(page) {
-    await page.getByRole('button', { name: 'Reiniciar' }).click();
-
-    const dialog = page.getByRole('dialog');
-    await expect(dialog.getByText('Reiniciar diagrama')).toBeVisible();
-
-    await dialog.getByRole('button', { name: 'Aceptar' }).click();
-}
-
-async function importDiagram(page, diagram) {
-    await page.getByRole('button', { name: 'Importar JSON' }).click();
-
-    const dialog = page.getByRole('dialog');
-    await expect(
-        dialog.getByText('Importación de diagrama desde JSON'),
-    ).toBeVisible();
-
-    const input = dialog.locator('input[type="file"]');
-
-    await input.setInputFiles({
-        name: 'diagram.json',
-        mimeType: 'application/json',
-        buffer: Buffer.from(JSON.stringify(diagram), 'utf8'),
-    });
-
-    await expect(
-        page.getByText('Diagrama importado con éxito.').last(),
-    ).toBeVisible();
-}
+import {
+    exportCurrentDiagram,
+    importDiagram,
+    resetDiagram,
+    seedSavedDiagram,
+} from '../helpers/persistence';
 
 test('relation configuration persists after accept and survives reload', async ({ page }) => {
     await page.goto('/');
@@ -154,14 +104,7 @@ test('reload preserves explicit attribute edge ids', async ({ page }) => {
         relations: [],
     };
 
-    // Seed a persisted diagram whose attribute-edge ids are explicit and non-arithmetic.
-    // The editor must preserve them after recreate + sync.
-    await page.addInitScript((initialDiagram) => {
-        window.localStorage.setItem(
-            'diagramData',
-            JSON.stringify(initialDiagram),
-        );
-    }, diagram);
+    await seedSavedDiagram(page, diagram);
 
     await page.goto('/');
 
@@ -254,12 +197,7 @@ test('export/import round-trip preserves diagram structure', async ({ page }) =>
         ],
     };
 
-    await page.addInitScript((initialDiagram) => {
-        window.localStorage.setItem(
-            'diagramData',
-            JSON.stringify(initialDiagram),
-        );
-    }, diagram);
+    await seedSavedDiagram(page, diagram);
 
     await page.goto('/');
